Fix exposure logging + cleanup
Logging conditional on whether there is a matching variable can cause
downstream (errors in
randomization)[https://github.com/HubSpot/react-experiments/issues/1].
This diff fixes that by not making logging dependent on having a
matching variant.  This diff also includes a number of changes that
increase specificity, including:

- Not making logging depend on whether there is a matching variant.
- Renaming variables to better between variant names, exposures, and
selected components
- Renaming Experiment to VariationExperiment, since this component
reflects a more limited type of experiment (a simple A/B test, or
"variation experiment") that people may want to run. This also helps
distinguish itself from PlanOut.js' experiment class
- Improvements to index.html: have exposure logging print to console
for easier debugging / understanding; using lowercase variable name to
refer to a class instance
- Added more specific development setup instructions -- it seemed that
one has to do an npm install for babel and related libraries for things
to run
- The web pack build was pointing to ../dist, so that running sh
build.sh would build to a directory above the repo. This makes it point
to ./dist.

Todo:
- Nuke utils.js
- Possibly get rid of experimentEnrollment. It seems to add a lot of
clutter and is mostly there for doing conditional logging.
- Rename experimentClass prop to experiment, since it expects an
instance of an experiment class
- In the future, we should include a legit propType for the
VariationExperiment class that corresponds to a PlanOut.Experiment
object (I tried doing this before but it didn't work)
- Mostly neglected the namespace class because it sounds like something
that Guy is going to nuke.